testing SingularityCheckingUpstreamsUpdate
