+++ SingularityService/src/main/java/com/hubspot/singularity/hooks/LoadBalancerClient.java
@@ -13,7 +13,7 @@ import com.hubspot.singularity.SingularityDeploy;
 import com.hubspot.singularity.SingularityLoadBalancerUpdate;
 import com.hubspot.singularity.SingularityRequest;
 import com.hubspot.singularity.SingularityTask;
-import com.hubspot.singularity.hooks.LoadBalancerClientImpl.SingularityCheckingUpstreamsUpdate;
+import com.hubspot.singularity.SingularityCheckingUpstreamsUpdate;
 
 public interface LoadBalancerClient {
